+++ tests/testthat/test-spex-extent.R
@@ -13,13 +13,13 @@ test_that("by id is not implemented", {
 
 test_that("extent of anything", {
   library(raster)
-  raster::extent(psf)
   expect_equal(raster::xmin(spex(psf, "+")), 0)
   expect_equal(raster::xmax(spex(psf, "+")), 1.23)
   
-  p <- par()
   expect_equal(raster::extent(spex()) , raster::extent(0, 1, 0, 1))
-  par(p)
+  
+  expect_equal(xlim(ice), c(-3950000,  3950000))
+  expect_equal(ylim(ice), c(-3950000,  4350000))
 })
 
 library(raster)