+++ NEWS.md
@@ -1,3 +1,15 @@
+# spex 0.5.0
+
+* New latitude functions `latitudecircle` to build a polar circle. 
+
+* Now return all cells as polygons if raster has no values. 
+
+* Improved handling of `na.rm`, also now TRUE by default. #13 Thanks to Adriano Fantini.
+
+* A modest speed-up (up to 2X) for `polygonize` by more careful use of R. 
+
+* better column name for resulting Spatial data frame
+
 # spex 0.4.0
 
 * new functions `xlim` and `ylim`, supporting anything understandable by `spex`